[C-IRIS] move cspace_free_polytope_test.cc out of dev folder. Also move several functions in dev/cspace_free_polytope.cc to out of dev.
