+++ geometry/optimization/dev/test/cspace_free_polytope_with_mosek_test.cc
@@ -7,7 +7,6 @@
 #include "drake/geometry/collision_filter_declaration.h"
 #include "drake/geometry/geometry_ids.h"
 #include "drake/geometry/optimization/cspace_free_polytope.h"  // NOLINT
-#include "drake/geometry/optimization/dev/test/c_iris_test_utilities.h"
 #include "drake/geometry/optimization/test/c_iris_test_utilities.h"
 #include "drake/multibody/rational/rational_forward_kinematics.h"
 #include "drake/multibody/rational/rational_forward_kinematics_internal.h"
